refactor: split utils.smAPI into two modules
