+++ metaspace/graphql/src/utils/index.ts
@@ -1,5 +1,5 @@
 export {createConnection, findUserByEmail} from './db';
-export {smAPIRequest} from './smAPI';
+export {smApiDatasetRequest} from './smApi/datasets';
 
 export type LooselyCompatible<T> =
   {[K in keyof T]?: T[K] extends (string | number | null | undefined) ? T[K] | null : any};
